pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "superhealth"
version = "1.0.0"
description = "Personal health data management system with LLM-powered insights"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "chasezjj", email = "chasezjj@163.com"},
]
keywords = ["health", "garmin", "fitness", "llm", "sqlite", "dashboard"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]
dependencies = [
    "pydantic>=2.0.0,<3.0",
    "requests>=2.31.0,<3.0",
    "pandas>=2.0.0,<3.0",
    "plotly>=5.20.0,<6.0",
    "streamlit>=1.32.0,<1.41.0",
    "flask>=3.0.0,<4.0",
    "scikit-learn>=1.4.0,<2.0",
    "python-dateutil>=2.8.0,<3.0",
    "reportlab>=4.0.0,<5.0",
    "kaleido>=0.2.1,<1.0",
    "exchangelib>=5.0.0,<6.0",
    "tomli>=2.0.0,<3.0; python_version < '3.11'",
    "tomli-w>=1.0.0,<2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]
garmin = ["playwright>=1.44.0,<2.0"]
claude = ["anthropic>=0.20.0"]
baichuan = ["openai>=1.0.0,<2.0"]
all = [
    "playwright>=1.44.0,<2.0",
    "anthropic>=0.20.0",
    "openai>=1.0.0,<2.0",
]

[project.scripts]
superhealth-pipeline = "superhealth.daily_pipeline:main"
superhealth-goals = "superhealth.goals.cli:main"

[project.urls]
Homepage = "https://github.com/chasezjj/superhealth"
Repository = "https://github.com/chasezjj/superhealth"
Issues = "https://github.com/chasezjj/superhealth/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --cov=superhealth --cov-report=term-missing --cov-fail-under=25"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501", "E402"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
